# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sales System API",
    "description": "نظام المبيعات المتقدم — API شامل لإدارة الحسابات والمعاملات المالية والموظفين والرواتب",
    "version": "1.0.0",
    "contact": { "name": "Sales System Team" }
  },
  "servers": [
    { "url": "/api", "description": "API Base" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token من POST /v1/auth/login"
      }
    },
...
  },
  "security": [{ "BearerAuth": [] }],
...
  "paths": {
    "/v1/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "تسجيل الدخول",
        "description": "التحقق من بيانات المستخدم وإرجاع JWT token",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string", "example": "admin" },
                  "password": { "type": "string", "example": "1234" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "نجاح — يحتوي على token وبيانات المستخدم" },
          "401": { "description": "بيانات دخول غير صحيحة" },
          "403": { "description": "حساب غير نشط" },
          "429": { "description": "حساب مقفول بسبب محاولات متعددة" }
        }
      }
    },
...
    "/v1/auth/create-user": {
      "post": {
        "tags": ["Auth"],
        "summary": "إنشاء مستخدم جديد (مدير النظام فقط)",
        "responses": {
          "201": { "description": "تم إنشاء المستخدم بنجاح" },
          "403": { "description": "غير مصرح — مطلوب صلاحيات مدير النظام" }
        }
      }
    },
    "/v1/accounts": {
      "get": {
        "tags": ["Accounts"],
        "summary": "عرض قائمة الحسابات",
        "responses": {
          "200": { "description": "قائمة الحسابات" }
        }
      },
      "post": {
        "tags": ["Accounts"],
        "summary": "إنشاء حساب جديد (admin فقط)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "maxLength": 100 },
                  "balance": { "type": "number", "minimum": 0 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "تم إنشاء الحساب بنجاح" },
          "400": { "description": "اسم الحساب مطلوب" }
        }
      }
    },
...
    "/v1/transactions": {
      "get": {
        "tags": ["Transactions"],
        "summary": "عرض قائمة المعاملات",
        "parameters": [
          { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["invoice", "payment", "purchase", "refund", "adjustment", "transfer", "fee", "settlement"] } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "direction", "in": "query", "schema": { "type": "string", "enum": ["debit", "credit"] } },
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 25 } }
        ],
        "responses": {
          "200": { "description": "قائمة المعاملات مع pagination" }
        }
      },
      "post": {
        "tags": ["Transactions"],
        "summary": "إنشاء معاملة مالية جديدة",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["type", "amount", "direction"],
                "properties": {
                  "type": { "type": "string", "enum": ["invoice", "payment", "purchase", "refund", "adjustment", "transfer", "fee", "settlement"] },
                  "amount": { "type": "number", "minimum": 0 },
                  "direction": { "type": "string", "enum": ["debit", "credit"] },
                  "description": { "type": "string" },
                  "paymentMethod": { "type": "string", "enum": ["cash", "check", "bank_transfer", "card", "online", "invoice", "other"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "تم إنشاء المعاملة بنجاح" },
          "400": { "description": "بيانات ناقصة أو غير صالحة" }
        }
      }
    },
...
    "/v1/users": {
      "get": {
        "tags": ["Users"],
        "summary": "عرض قائمة المستخدمين",
        "parameters": [
          { "name": "role", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } }
        ],
        "responses": {
          "200": { "description": "قائمة المستخدمين مع pagination" }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "إنشاء مستخدم جديد (admin فقط)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password", "name"],
                "properties": {
                  "username": { "type": "string", "maxLength": 50 },
                  "password": { "type": "string", "minLength": 6, "maxLength": 128 },
                  "name": { "type": "string", "maxLength": 100 },
                  "email": { "type": "string", "format": "email" },
                  "role": { "type": "string", "enum": ["system_admin", "admin", "manager", "user", "sales", "inventory"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "تم إنشاء المستخدم بنجاح" },
          "409": { "description": "اسم المستخدم موجود بالفعل" }
        }
      }
    },
...
    "/v1/sync/broadcast": {
      "post": {
        "tags": ["Sync"],
        "summary": "بث إشعار تغيير لجميع العملاء المتصلين (admin فقط)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": { "type": "string" },
                  "section": { "type": "string" },
                  "data": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "تم البث بنجاح" }
        }
      }
    },
    "/v1/backups": {
      "get": {
        "tags": ["Backups"],
        "summary": "عرض النسخ الاحتياطية (آخر 20)",
        "responses": {
          "200": { "description": "قائمة النسخ الاحتياطية" }
        }
      },
      "post": {
        "tags": ["Backups"],
        "summary": "إنشاء نسخة احتياطية جديدة",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["data"],
                "properties": {
                  "data": { "type": "object", "description": "البيانات المطلوب نسخها احتياطياً" },
                  "label": { "type": "string", "maxLength": 100 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "تم إنشاء النسخة الاحتياطية" }
        }
      }
    },
...
    "/v1/company": {
      "get": {
        "tags": ["Company"],
        "summary": "عرض بيانات الشركة",
        "responses": {
          "200": { "description": "بيانات الشركة" }
        }
      },
      "patch": {
        "tags": ["Company"],
        "summary": "تحديث إعدادات الشركة (admin فقط)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "industry": { "type": "string" },
                  "country": { "type": "string" },
                  "status": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "تم التحديث بنجاح" }
        }
      }
    },
...
  }
}